Back-fit reference features to a subset of spectra

Back-fit reference features to a subset of spectra

backfit_ref.feats.2.subset.specs(
  m.inds,
  fits.feature,
  match.info,
  feature,
  xmat,
  ppm,
  plots = T
)

Arguments

m.inds

A vector of indices corresponding to the matches between reference features and spectra. Necessary to allow for subsetting of matches (i.e. because of filtering)

fits.feature

A list of fitted features (to reference spectra)

match.info

A data frame containing information about the matches between features and reference spectra

feature

List containing feature intensities and positions

xmat

Full spectral matrix from which the features were derived

ppm

A vector containing the ppm axis of the spectra

plots

A logical value indicating whether to generate plots of the fits and back-fit feasibility scores - only use for a couple of backfits at a time as plots will double the weight of the result

Value

A list containing the back-fits and back-fit feasibility scores for each spectrum in the subset, as well as a grid plot of the fits and back-fit feasibility scores.