corrPocketPairs

corrPocketPairs

corrPocketPairs_al(
  x,
  ppm,
  ws,
  reg = NULL,
  plotHeatmap = FALSE,
  wdlimit = 0.99,
  rcutoff = 0.75
)

Arguments

x

a matrix with numerical values.

ppm

a numeric vector of the same length as ncol(x) specifying the ppm for each column.

ws

an integer specifying half the size of the sliding window used to calculate correlations.

reg

an optional vector specifying the column indices to consider in x.

plotHeatmap

a logical indicating whether to plot a heatmap of the correlation matrix.

wdlimit

a numeric (0,1) specifying the minimum fraction of sliding windows that need to contain a pocket to consider a point noise.

rcutoff

a numeric (0,1) specifying the correlation coefficient cutoff to use when extracting significant peaks.

Value

a list with the following elements:

regions

a matrix specifying the indices of the peaks in each column of the original matrix, filtered to exclude peaks that are too small or only have unidirectional interactions.

corr

a matrix of correlation coefficients between columns of x.

cov

a matrix of covariance values between columns of x.

peakMap

a matrix specifying the indices of the peaks in each column of the correlation matrix.

noiseDist

the proportion of pockets containing each window index.