Expand a correlation window around a given peak index to the nearest correlation minima outside of the window bounds.

Expand a correlation window around a given peak index to the nearest correlation minima outside of the window bounds.

corr_expand_window(xmat, ppm, peakInd, wind)

Arguments

xmat

A matrix of NMR spectra.

ppm

A vector of ppm values corresponding to the columns of xmat.

peakInd

The index of the peak around which to expand the window.

wind

A vector defining the window (column inds) around the peak.

Value

A list with the following elements:

peak

The index of the peak used for window expansion.

intcorr

The integral of the correlation within the expanded window.

data

A data frame containing the correlation values, covariance values, and corresponding ppm values within the expanded window.

sr

The STOCSY object for the expanded window.

wind

The original window vector.

newWind

The expanded window vector.

corrRbound

The index of the right boundary of the expanded window in the wind vector.

corrLbound

The index of the left boundary of the expanded window in the wind vector.