Match a feature against a reference using FFT-based convolution and correlation picks the top max.hits peaks in the convolution, then calculates correlation and pvalue of the PCC at those points in the ref. Returns a dataframe with the xcorr information.

Match a feature against a reference using FFT-based convolution and correlation picks the top max.hits peaks in the convolution, then calculates correlation and pvalue of the PCC at those points in the ref. Returns a dataframe with the xcorr information.

feature_match2ref_slim(
  f.num,
  r.num,
  feat,
  ref,
  feat.ft.c,
  ref.ft,
  pad.size,
  max.hits = 5,
  r.thresh = 0.8,
  p.thresh = 0.01
)

Arguments

f.num

numeric: Feature number

r.num

numeric: Reference number

feat

numeric: Vector of feature intensities

ref

numeric: Vector of reference intensities

feat.ft.c

numeric: FFT of feature intensities, conjugated

ref.ft

numeric: FFT of reference intensities

pad.size

numeric: Size of the padding applied during FFT-based convolution

max.hits

numeric: Maximum number of candidate lags to consider

r.thresh

numeric: Correlation threshold for considering a match

p.thresh

numeric: P-value threshold for considering a match

Value

A data frame with the following columns: - feat: Feature number or identifier - ref: Reference number or identifier - lag: The lag that produces the highest correlation - rval: The correlation coefficient - pval: The p-value of the correlation - pts.matched: Number of data points used in the correlation - pts.feat: Number of data points in the feature vector - feat.start: The starting index of the feature vector used in the correlation - feat.end: The ending index of the feature vector used in the correlation - ref.start: The starting index of the reference vector used in the correlation - ref.end: The ending index of the reference vector used in the correlation