Filter features based on specified criteria - null features - features outside of ppm range - features with no runs >= min.runlength - features derived from < min.subset spectra - features with strong baseline effect

Filter features based on specified criteria - null features - features outside of ppm range - features with no runs >= min.runlength - features derived from < min.subset spectra - features with strong baseline effect

filterFeatures(
  feature,
  ppm,
  ppm.range,
  min.runlength = 3,
  min.subset = 5,
  prom.ratio = 0.3,
  give = "filter"
)

Arguments

feature

A feature object to be filtered

ppm

A vector of ppm values for each point in the spectra

ppm.range

A vector specifying the range of ppm values to include

min.runlength

The minimum length of a run of consecutive data points to include

min.subset

The minimum number of spectra a feature must be present in to be included

prom.ratio

The maximum ratio of peak prominence to intensity range for a feature to be considered monotonic

give

A character string indicating whether to return the filtered feature object or the filter as a logical vector

Value

If give = "features", a feature object with the filtered features. If give = "filter", a logical vector indicating which features passed the filter.