Pad matrix with rows and columns of specified value

This function takes a matrix and pads it with rows and columns of a specified value. The number of rows and columns to add can also be specified. The new rows and columns will have the same value.

padmat(x, use = NA, row.by = 0, col.by = 1)

Arguments

x

A matrix to pad

use

The value to fill new rows and columns with (default is NA)

row.by

The number of rows to add to each side of the matrix

col.by

The number of columns to add to each side of the matrix

Value

A padded matrix

See also

Examples

m <- matrix(1:4, ncol = 2)
padmat(m, use = 0, row.by = 1, col.by = 1)
#> Error in padmat(m, use = 0, row.by = 1, col.by = 1): could not find function "padmat"

m2 <- matrix(1:6, ncol = 2)
padmat(m2, row.by = 2, col.by = 3)
#> Error in padmat(m2, row.by = 2, col.by = 3): could not find function "padmat"

m3 <- matrix(1:9, ncol = 3)
padmat(m3, use = -1, row.by = 1, col.by = 2)
#> Error in padmat(m3, use = -1, row.by = 1, col.by = 2): could not find function "padmat"