Plot STOCSY line over the current plot

Plot STOCSY line over the current plot

plot_addSTOCSYLine(
  g,
  specRegion,
  ppm,
  corr = NULL,
  covar = NULL,
  driver = NULL,
  bounds = NULL
)

Arguments

g

plot to add STOCSY line to

specRegion

region of the NMR spectrum to plot

ppm

vector of chemical shift values

corr

matrix of correlations between reference and query spectra

covar

matrix of covariances between reference and query spectra

driver

chemical shift value of the driver peak, if given

bounds

chemical shift values of the boundaries of the region of interest, if given

Value

plot with STOCSY line added