Plot the STORM results

Plot the storm results (s) from storm_play, etc. Plot light gray ref region for optimal subset. use in lapply(). Pass in s: nothing outside

plot_storm_refRegions(
  xmat,
  ppm,
  s,
  bgplot = "overlayed",
  vshift = 1,
  hshift = 0.05,
  calcStocsy = TRUE,
  n_xticks = 4
)

Arguments

xmat

numeric matrix of spectral data

ppm

numeric vector of length equal to ncol(xmat)

s

STORM results from storm_play or similar function

bgplot

a character vector indicating how the plot is rendered: "overlayed" or "stack". Default is "overlayed".

vshift

numeric scalar indicating the vertical shift between spectra when bgplot is "stack". Default is 1.

hshift

numeric scalar indicating the horizontal shift between spectra when bgplot is "stack". Default is 0.05.

calcStocsy

a logical indicating whether to calculate STOCSY plot. Default is TRUE.

n_xticks

an integer indicating the number of ticks on the x-axis. Default is 4.

Value

a ggplot object with the plot of the spectral data, the reference region, and the STOCSY plot (if applicable).

See also

Examples

plot_storm_refRegions(xmat = xmat, ppm = ppm, s = stormResults[[1]])
#> Error in plot_storm_refRegions(xmat = xmat, ppm = ppm, s = stormResults[[1]]): could not find function "plot_storm_refRegions"