Prepare reference spectra for matching to a dataset To match features from a dataset to reference spectra in fine detail, they need to have the same spectral axis, more or less. This function interpolates each reference spectrum in the list to the dataset ppm axis, and then filters its signal to avoid inclusion of noise/zeros, which are replaced with NAs so that cross-correlations, etc. are minimized. In future versions, a solvent region could be provided for each (or all) that will be replaced with NAs to avoid matching that peak.

Takes gissmo (or other) reference spectra as a list with - tag (short name, e.g. "bmse000005" - not necessarily unique) - ref.name (should be unique with field strength etc.) - ppm axis - ref spectrum data

prepRefs.for.dataset(data.list, ppm.dataset, ref.sig.SD.cutoff = 0.01)

Arguments

data.list

a list of reference spectra, where each element of the list is a list containing the fields "tag", "ref.name", "compound.name", "ppm", and "data"

ppm.dataset

the ppm axis of the dataset to which the reference spectra will be matched

ref.sig.SD.cutoff

the cutoff for reference spectra signal, expressed as a fraction of the standard deviation

Value

a list of processed reference spectra, where each element of the list is a list containing the fields "tag", "ref.name", "compound.name", "ppm", "data", and "mapped". The "mapped" field contains a list with the fields "ppm", "data", "sig.cutoff", and "sd.cutoff"

Examples

data.list <- list(ref1, ref2, ref3)
#> Error in eval(expr, envir, enclos): object 'ref1' not found
ppm.dataset <- c(0, 1, 2, 3, 4)
ref.sig.SD.cutoff <- 0.01
prepRefs.for.dataset(data.list, ppm.dataset, ref.sig.SD.cutoff)
#> Error in prepRefs.for.dataset(data.list, ppm.dataset, ref.sig.SD.cutoff): could not find function "prepRefs.for.dataset"