Calculate sliding window correlation matrix, and label correlation peaks (pockets)

Calculate sliding window correlation matrix, and label correlation peaks (pockets)

slidingCorr(
  x,
  ws,
  extractPockets = FALSE,
  plotting = TRUE,
  vshift = 20,
  ppm = NULL
)

Arguments

x

Input matrix.

ws

Size of sliding window.

extractPockets

If TRUE, calculate correlation pockets for each column of x within each sliding window.

plotting

If TRUE, produce a stackplot of the correlation matrix with pockets highlighted.

vshift

Vertical shift in plot.

ppm

X-axis values for plot. If not specified, uses column numbers of x.

Value

A list containing the following items:

corr_compact

The compact form of the sliding correlation matrix.

indsmat

The offset matrix used to calculate the sliding correlation matrix.

cov_compact

The compact form of the sliding covariance matrix.

isPocket

A logical matrix indicating whether each point in the sliding window is within a correlation pocket.

plot

If plotting is TRUE, a stackplot of the correlation matrix. If extractPockets = T, show only the pockets.

window

The relative inds in the sliding window.

center

The center position of the sliding window.

Examples

# Generate a test matrix
x <- matrix(rnorm(2000), nrow = 100)

# Calculate sliding correlation matrix with pockets and plot
slidingCorr(x, ws = 20, extractPockets = TRUE, plotting = TRUE)
#> Error in slidingCorr(x, ws = 20, extractPockets = TRUE, plotting = TRUE): could not find function "slidingCorr"