Plot stacked spectra with ggplot2 and ggridges

stackplot(
  ymat = NULL,
  xvect = NULL,
  vshift = 1,
  hshift = 0,
  shift_by = NULL,
  xdir = "reverse",
  show = TRUE
)

Arguments

ymat

A numeric matrix of spectral intensities.

xvect

A numeric vector of ppm values.

vshift

A numeric value for shifting spectra vertically (spectra are scaled down to this value).

hshift

A numeric value for shifting spectra horizontally (as

shift_byA numeric vector for shifting spectra by row number.

xdirA character string for x-axis direction, either "reverse" or "forward".

showA logical value for whether or not to display the plot.

A ggplot2 object.

This function plots stacked spectra using ggplot2 and ggridges, which creates a density plot of the spectra that can help to visualize overlaps and peaks.

data(metab) stackplot(metab[, 1:10], xvect = metab[, 11])