storm_pairplay: Run modified STORM on the provided spectral region and ref shape. Built for accepting corrPocketPairs results. Notes:

STORM: Joram Posma's STORM has been adapted and optimized to accept these

storm_pairplay(
  xmat = NULL,
  ppm = NULL,
  b = 30,
  corrthresh = 0.8,
  q = 0.05,
  minpeak = 10,
  refSpec,
  ref.idx
)

Arguments

xmat

A matrix of spectral data (rows are spectra, columns are spectral points)

ppm

A vector of the spectral points in ppm (optional, default is all columns of xmat)

b

An integer giving the expansion parameter for the reference peak

corrthresh

A numeric giving the minimum correlation value to be considered for inclusion (for both subset AND reference optimization)

q

A numeric giving the p-value threshold for correlation significance (both subset AND reference optimization)

minpeak

An integer giving the minimum number of points allowed in a run of significant points in the reference

refSpec

A vector of spectral data to use as the initial reference

ref.idx

A vector of the spectral points (columns of xmat) to use as the initial reference

Value

A list with components "reconstructed" and "status". "reconstructed" is a matrix containing the reconstructed metabolite concentrations (rows are samples, columns are metabolites). "status" is a character string indicating whether the method converged successfully or failed.