Calculate cross-correlation between reference and shifted spectra

Wrote my own cross-correlation function for more consistent behavior in the context of comparing ref signatures to spec data. This takes a (wider) spec vector and slides it across the ref signature lags. The lags are used to extract a window of spec for each shift (lag). These windows are collected as rows of a matrix, which is then used to compute all correlations to the ref. MTJ 2023

xcorr(spec, ref.vals, ref.idx, lags, min.overlap = 3)

Arguments

spec

Numeric matrix. A matrix of shifted spectra.

ref.vals

Numeric vector. The values of the reference spectrum.

ref.idx

Integer vector. The indices of the reference spectrum.

lags

Integer vector. The range of lags to consider.

min.overlap

Integer. The minimum number of overlapping data points between reference and shifted spectra to calculate correlation.

Value

A list containing the following elements:

rvals

Numeric vector. The correlation scores for each shifted spectrum.

pvals

Numeric vector. The p-values for each correlation score.

overlaps

Numeric vector. The number of overlapping data points between reference and each shifted spectrum.

not.used

Logical vector. Indicates which shifted spectra did not meet the minimum overlap requirement and were not used in calculating correlation scores or p-values.

lags

Integer vector. The lags used for each shifted spectrum.

bestFit

List. Contains information on the best-matching shifted spectrum. The list has the following elements: which.lagLogical vector. Indicates which lags were the best match. rNumeric. The correlation score for the best match. lagInteger. The lag corresponding to the best match. pvalNumeric. The p-value for the correlation score of the best match. overlapsInteger. The number of overlapping data points between the reference and the best-matching shifted spectrum. specInds.matchedInteger vector. The indices of the data points in the shifted spectrum that correspond to the overlapping region with the reference spectrum. This can be used to plot the overlapping regions.

Examples

# Generate sample data
set.seed(123)
ref <- rnorm(10)
ref.idx <- 1:10
spec <- matrix(rnorm(1000), nrow = 10)
lags <- -5:5

# Calculate cross-correlation
xcorr(spec, ref, ref.idx, lags)
#> Error in xcorr(spec, ref, ref.idx, lags): could not find function "xcorr"