Cross-correlation with local alignment

This function calculates the cross-correlation between a reference vector and a matrix of spectra, where each spectrum is a row of the matrix. The function aligns the reference vector with each spectrum by shifting the spectrum and then calculating the cross-correlation between the aligned spectrum and the reference vector. The function returns the best shift for each spectrum, along with the corresponding correlation score, p-value, and overlap size. The function can also generate a plot showing the aligned spectra and reference vector.

xcorr_localign(
  x,
  ppm,
  signature.vals,
  signature.idx.wind,
  currentInds,
  min.overlap = 3,
  slide = "internal",
  lag.limit = NA,
  plots = F
)

Arguments

x

A matrix where each row is a spectrum.

ppm

A vector of column names for the matrix x.

signature.vals

A vector of values matching the signature index window.

signature.idx.wind

The reference indices (across all rows of currentInds).

currentInds

The window(s) in which the reference sits in x (often larger than the reference).

min.overlap

The minimum number of points on which to base the correlation.

slide

If "internal", the function just slides the reference vector until it hits the bounds of currentInds; if "external", the function pads currentInds to ensure that all reference vector points cross the bounds of currentInds; if "limited", the user provides the number of lags (duplicated for both directions).

lag.limit

The limit on the number of lags to consider (for "limited" slide option).

plots

A logical value indicating whether to generate a plot showing the aligned spectra and reference vector.

Value

A list with the following elements:

best.shifts

A vector of the best shift for each spectrum.

best.rvals

A vector of the correlation scores for each spectrum.

best.pvals

A vector of the p-values for each spectrum.

shiftedInds

A matrix of the indices in x for the aligned spectra.

shiftedSpecs

A matrix of the aligned spectra.

ref.aligned

A vector of the reference vector values (aligned with the spectra).

overlap.sizes

A vector of the overlap sizes for each spectrum.

g

A plot showing the aligned spectra and reference vector (if plots = TRUE).

Examples

x <- matrix(rnorm(100), ncol = 10)
ppm <- seq(100)
signature.vals <- rnorm(10)
signature.idx.wind <- seq(5, 14)
currentInds <- matrix(seq(1, 100, by = 10), ncol = 2)
xcorr_localign(x, ppm, signature.vals, signature.idx.wind, currentInds, min.overlap = 3, slide = "internal", lag.limit = NA, plots = FALSE)
#> Error in xcorr_localign(x, ppm, signature.vals, signature.idx.wind, currentInds,     min.overlap = 3, slide = "internal", lag.limit = NA, plots = FALSE): could not find function "xcorr_localign"